crashDat(elrm)	R Documentation

Crash Dataset: Calibration of Crash Dummies in Automobile Safety Tests

Description

Data from 58 simulated car crashes were analyzed. The relationship between the crash outcome (fatal, non-fatal) and 3 covariates was modeled.

Usage

data(drugDat)

Format

Matrix with columns:

[,1] 	 age 	 numeric 	 designed age of the crash dummy

[,2] 	 vel 	 numeric 	 velocity of the car at impact

[,3] 	 acl 	 numeric 	 acceleration of the car at impact

[,4] 	 y 	 numeric 	 outcome from test (1=fatal, 0=non-fatal)

[,5] 	 n 	 numeric 	 number of binomial trials. For this data, n is a vector of ones, thus each trial is modeled as a Bernoulli random variable.

[Package elrm version 1.1.3 Index]