titanDat(elrm)	R Documentation

Titanic Dataset

Description

Relationship between survival and passenger class on the Titanic. The records of the sinking of the Titanic were studied to establish the relationship between survival and passenger class on the ship. For each person on board the ocean liner, this dataset records sex, age (child/adult), class (crew, 1st, 2nd, 3rd class) and whether or not the person survived.

Usage

data(titanDat)

Format

Matrix with columns:

[,1] 	 surv 	 numeric 	 number of survivors

[,2] 	 n 	 numeric 	 total number of people

[,3] 	 class 	 numeric 	 passenger class (0 = crew, 1 = first, 2 = second, 3 = third)

[,4] 	 age 	 numeric 	 age group (1 = adult, 0 = child))

[,5] 	 sex 	 numeric 	 gender (1 = male, 0 = female)

References

Mehta CR, Patel NR. Logxact for Windows. Cytel Software Corporation, Cambridge, USA, 1999.

[Package elrm version 1.1.3 Index]