utiDat {elrm}	R Documentation

Urinary Tract Infection and Contraceptive Use

Description

How is the development of first-time urinary tract infection (UTI) related to contraceptive use? A study of sexually active college women with UTI was conducted, and their use of various contraceptives was surveyed.

Usage

data(utiDat)

Format

Matrix with columns:

[,1] 	 uti 	 numeric 	 infection status (1=yes, 0=no)

[,2] 	 n 	 numeric 	 number of binomial trials

[,3] 	 age 	 numeric 	 age of the person

[,4] 	 current 	 numeric 	 no regular current sex partner (1=yes,0=no)

[,5] 	 dia 	 numeric 	 diaphragm use (1=yes, 0=no)

[,6] 	 oc 	 numeric 	 oral contraceptive (1=yes, 0=no)

[,7] 	 pastyr 	 numeric 	 no regular partner with relationship < 1yr (1=yes, 0=no)

[,8] 	 vi 	 numeric 	 vaginal intercourse (1=yes, 0=no)

[,9] 	 vic 	 numeric 	 vaginal intercourse with condom (1=yes, 0=no)

[,10] 	 vicl 	 numeric 	 vaginal intercourse with lubricated condom (1=yes, 0=no)

[,11] 	 vis 	 numeric 	 vaginal intercourse with spermicide (1=yes, 0=no)

References

Mehta CR, Patel NR. Logxact for Windows. Cytel Software Corporation, Cambridge, USA, 1999.

[Package elrm version 1.2.3 Index]