diabDat(elrm)	R Documentation

Simulated Diabetes Dataset

Description

This dataset was simulated by sampling cases from an existing type 1 diabetes study (the original data could not be disclosed). The study investigates the relationship between concentration levels (low and high) of the islet antigen 2 antibody (IA2A) and several covariates of potential interest in type 1 diabetes patients (age, gender, and the number of copies (0,1 or 2) of the HLA-DQ2, HLA-DQ8 and HLA-DQ6.2 haplotypes).

Usage

data(diabDat)

Format

Matrix with columns:

[,1] 	 n 	 numeric 	 number of binomial trials.

[,2] 	 IA2A 	 numeric 	 number of patients with high concentration levels of the islet antigen 2 antibody

[,3] 	 gender 	 numeric 	 gender of patient

[,4] 	 age 	 numeric 	 age of patient

[,5] 	 nDQ2 	 numeric 	 number of copies of the HLA-DQ2 haplotype (0,1, or 2)

[,6] 	 nDQ8 	 numeric 	 number of copies of the HLA-DQ8 haplotype (0,1, or 2)

[,7] 	 nDQ6.2 	 numeric 	 number of copies of the HLA-DQ6.2 haplotype (0,1, or 2)

[Package elrm version 1.1.3 Index]